application/pyproject.toml
[tool.poetry]
name = "fastapi-todo-rest"
version = "0.1.0"
description = ""
authors = ["golubev <951364+golubev@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "~1.10.4"
celery = "~5.2.7"
email-validator = "~1.3.1"
emails = "~0.6"
fastapi = "~0.95.1"
jinja2 = "~3.1.2"
passlib = {version = "~1.7.4", extras = ["bcrypt"]}
pydantic = "~1.10.7"
python-jose = {version = "~3.3.0", extras = ["cryptography"]}
python-multipart = "~0.0.6"
psycopg2-binary = "~2.9.6"
sqlalchemy = "~1.4.48"
uvicorn = "~0.22.0"

[tool.poetry.group.dev.dependencies]
autoflake = "~1.7.8"
black = "~23.3.0"
faker = "~18.7.0"
flake8 = "~5.0.4"
httpx = "~0.24.0"
isort = "~5.12.0"
mypy = "~1.2.0"
pytest = "~7.3.1"
pytest-cov = "~4.0.0"
sqlalchemy = {version = "~1.4.48", extras = ["mypy"]}
types-passlib = "~1.7.4"
types-python-jose = "~3.3.0"
celery-types = "~0.15.0"

[tool.black]
line-length = 88

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["src", "tests"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
extend_skip = ["alembic"]
